Move panel/panel.py to panel_view.py
