+++ commands.py
@@ -2,7 +2,7 @@ import sublime
 import sublime_plugin
 
 from .lint import persist
-from .panel.panel import fill_panel, PANEL_NAME
+from .panel_view import fill_panel, PANEL_NAME
 
 
 class SublimeLinterLintCommand(sublime_plugin.TextCommand):
