+++ status_bar_view.py
@@ -4,7 +4,7 @@ import sublime_plugin
 from .lint import persist
 from .lint.const import STATUS_KEY
 from .lint import events
-from .panel import panel
+from . import panel_view
 
 
 State = {
@@ -47,7 +47,7 @@ class UpdateState(sublime_plugin.EventListener):
             'we_count': we_count,
             'current_pos': current_pos
         })
-        panel.update_panel_selection(**State)
+        panel_view.update_panel_selection(**State)
         draw(**State)
 
     def on_selection_modified_async(self, _primary_view_):
@@ -57,7 +57,7 @@ class UpdateState(sublime_plugin.EventListener):
             State.update({
                 'current_pos': current_pos
             })
-            panel.update_panel_selection(**State)
+            panel_view.update_panel_selection(**State)
             draw(**State)
