+++ sublime_linter.py
@@ -15,7 +15,7 @@ from .lint import persist, util, style
 from .lint.error import ErrorStore
 from .lint.const import WARN_ERR
 from .lint import backend
-from .panel import panel
+from . import panel_view
 
 
 def backup_old_settings():
@@ -153,7 +153,7 @@ class Listener:
                 d.pop(vid, None)
 
         queue.cleanup(vid)
-        panel.fill_panel(view.window(), update=True)
+        panel_view.fill_panel(view.window(), update=True)
 
     def on_hover(self, view, point, hover_zone):
         """On mouse hover event hook.
@@ -267,7 +267,7 @@ class SublimeLinter(sublime_plugin.EventListener, Listener):
             highlights.draw(view)
 
         for window in sublime.windows():
-            panel.fill_panel(window, update=True)
+            panel_view.fill_panel(window, update=True)
 
     def hit(self, view):
         """Record an activity that could trigger a lint and enqueue a desire to lint."""
